+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.2 under development
 
 - Chg #297: Remove `QueryBuilder::getColumnType()` child method as legacy code (@Tigrov)
+- Enh #300: Refactor insert default values (@Tigrov)
 
 ## 1.0.1 July 24, 2023